create last two app components
