+++ src/app/components/mod.rs
@@ -2,6 +2,8 @@ pub mod footer;
 pub mod library_component;
 pub mod menu_bar;
 pub mod player_component;
+pub mod playlist_table;
+pub mod playlist_tabs;
 
 pub trait AppComponent {
     type Context;
